+++ UWP/ViewModels/PortfolioViewModel.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
+using Microsoft.Toolkit.Mvvm.Messaging;
+using System;
 using System.Collections.ObjectModel;
 using UWP.Models;
 
@@ -72,6 +74,12 @@ namespace UWP.ViewModels {
 		public bool AllPurchasesInCurrency {
 			get => allPurchasesInCurrency;
 			set => SetProperty(ref allPurchasesInCurrency, value);
+		}
+
+		/// #############################################################################
+		public void InAppNotification(string title, string message = "") {
+			var tuple = new Tuple<string, string>(title, message);
+			Messenger.Send(new NotificationMessage(tuple));
 		}
 	}
 }
